requirements.txt
aio-pika==6.8.0
aiofiles==0.7.0
aiohttp==3.8.0
aiormq==3.3.1
aiosignal==1.2.0
anyio==3.3.4
asgiref==3.4.1
asn1crypto==1.4.0
astroid==2.8.4
async-timeout==4.0.0
attrs==21.2.0
boto3==1.20.8
botocore==1.23.8
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
fastapi==0.70.0
frozenlist==1.2.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
isort==5.10.0
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
lxml==4.6.4
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
pamqp==2.3.0
pg8000==1.22.0
pika==1.2.0
platformdirs==2.4.0
pydantic==1.8.2
pylint==2.11.1
python-dateutil==2.8.2
python-dotenv==0.19.1
requests==2.26.0
s3transfer==0.5.0
scramp==1.4.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
toml==0.10.2
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
wrapt==1.13.3
xmltodict==0.12.0
yarl==1.7.2
